move model construction out of codegen
